% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/all.f.R
\name{all.f}
\alias{all.f}
\title{Plots a base boxplot with significant multiple comparisons.}
\usage{
\method{all}{f}(x, y, ylab = "")
}
\arguments{
\item{x}{The grouping variable (of a 'data' data frame) for the box plot
as character to fit scheme: data[,x]}

\item{y}{The outcome variable (of a 'data' data frame) for the box plot
as character to fit scheme: data[,y]}

\item{ylab}{The label of the y axis}
}
\value{
The plot with significant multiple comparisons.
}
\description{
Plot for x- grouping variable and y - continous variable with lines
illustrating significant differences among groups in multiple comparisons
with Dunn's test.
}
\examples{
data<-mtcars
data$cyl<-as.factor(data$cyl)
all.f("cyl","mpg")
}

