\name{anareg-package}
\alias{anareg-package}
\alias{anareg}
\docType{package}
\title{
This package contains functions to evaluate the Anaphylaxis Registry.
}
\description{
Anaphylaxis Registry is maintained by the Allergy Centrum Charite in Berlin. 
This package provides functions to munging and visualization of its database.
}
\details{
\tabular{ll}{
Package: \tab anareg\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2015-04-04\cr
License: \tab GPL-2\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Wojciech Francuzik

Maintainer: Wojciech Francuzik <wojciech.francuzik@charite.de>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
%\examples{
%}
