% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AutomaticOR.proc.R
\name{AutomaticOR.proc}
\alias{AutomaticOR.proc}
\title{Makes a list of odds ratios between levels of two variables.}
\usage{
AutomaticOR.proc(x, y, alpha = 0.05)
}
\arguments{
\item{x}{The model (formal class) for regression for which to plot odds
ratios}
}
\value{
The plot of odds ratios
}
\description{
It takes a model of a lgistic, linear, ordinal, regression and plots the odds
ratios where
}
\examples{
plot_odds(regression.model)
}

