% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cc.restrictor.R
\name{cc.restrictor}
\alias{cc.restrictor}
\title{Make complete cases data base for AIC analysis.}
\usage{
cc.restrictor(outcome, predictors, db, single.set = F)
}
\arguments{
\item{outcome}{is the character string specifying the outcome variable.
This variable has to be a part of the database (db)}

\item{predictors}{Predictors are the known variables, based on which
we want to conduct our predicting.}

\item{db}{The database on which we should perform the operations}

\item{single.set}{is a logical if TRUE then the function will provide only
one training set, if FALSE if will return both training set and a test set.}
}
\description{
This function makes a list with three values FORMULA, TRAIN data set and
TEST data set.  Very useful for regression modelling.
}
\examples{
complete.cases <- cc.restrictor("ANAscore",c("d_age","b_sex"),data)
}

