% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/m_chi.R
\name{m_chi}
\alias{m_chi}
\title{Quickly measure multiple chi sq tests out of many variables}
\usage{
m_chi(v, data)
}
\arguments{
\item{v}{The vector of character variable names or numbers corresponding to
the said variables}

\item{data}{Data frame where the variables are to be calculated.}
}
\value{
The DATA FRAME of chisq.test values in every configuration of given
set of variables
}
\description{
It takes a vector of numbers referred to the variables you want to check
and prints out the chi test values in a data frame.
}
\examples{
multi.chis <- m_chi(c("d_age","b_sex"),data)
}

