% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.anascore.R
\name{make.anascore}
\alias{make.anascore}
\title{Makes ANAscore variable with the given set of predefined variables}
\usage{
make.anascore(symptoms = c(17:21, 23:29, 31:39, 41:55, 65), weights = c(3,
  2, 1, 2, 2, 3, 1, 4, 3, 4, 12, 3, 8, 4, 1, 20, 4, 2, 8, 8, 1, 20, 18, 20, 10,
  6, 18, 12, 10, 18, 2, 8, 3, 15, 20, 6, 50), grades = c(1, 1, 1, 1, 1, 1, 1,
  2, 1, 2, 3, 1, 2, 2, 1, 4, 2, 1, 2, 2, 1, 4, 3, 4, 2, 2, 3, 3, 2, 3, 1, 2, 1,
  3, 4, 2, 5), data = data)
}
\arguments{
\item{symptoms}{The vector of character variable names or numbers
corresponding to the variables in correct order corresponding to proper
weights.}

\item{weights}{The vector of the same lenght as symptoms, defining the
impact symptom have on the severity of anaphylaxis.}

\item{grades}{The vector of the same lenght as symptoms, defining the
severity grades of the symptoms. Try not to make more than 5 grades.}

\item{data}{Data frame where the variables are to be calculated.}
}
\value{
The vector of ANAscore values.
}
\description{
It takes a vector of numbers referred to the variables you want to include.
The variables have to be in the correct order to properly assign weights.
the
}
\examples{
data$ANAscore <- make.anascore()
}

