% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.na.R
\name{make.na}
\alias{make.na}
\title{Deletes the "UNKNOWN" variable levels from the factor.}
\usage{
make.na(x)
}
\arguments{
\item{x}{The value to check if "Unknown" is the in the variable levels}
}
\value{
The changed value with NA insted of "unknown"
}
\description{
It takes a vector of numbers referred to the variables you want to include.
The variables have to be in the correct order to properly assign weights.
the
}
\examples{
sapply(x.vector,make.na)
}

