% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_odds.R
\name{plot_odds}
\alias{plot_odds}
\title{Makes a plot of odds ratios from a regression model.}
\usage{
plot_odds(x, title = NULL)
}
\arguments{
\item{x}{The model (formal class) for regression for which to plot odds
ratios}
}
\value{
The plot of odds ratios
}
\description{
It takes a model of a lgistic, linear, ordinal, regression and plots the odds
ratios where
}
\examples{
plot_odds(regression.model)
}

