% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/show.odds.R
\name{show.odds}
\alias{show.odds}
\title{Calculates the odds ratios for the predictors of regression models.}
\usage{
show.odds(model, digits)
}
\arguments{
\item{model}{The regression class model.}

\item{digits}{The number of digits after 0 to print in the data frame.}
}
\value{
The table of odds ratios with confidence intervals and p values, in
character format for easier printing by the xtable.
}
\description{
This function plots a table of odds ratios with 95% CI and p values for the
predictors in a regression model.
}
\examples{
xtable(which.restricts(c("d_age","b_sex"),"ANAscore",data))
}

