% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/which.restricts.R
\name{which.restricts}
\alias{which.restricts}
\title{Tells which variable to eliminate to get more observations.}
\usage{
which.restricts(predictors, outcome, db)
}
\arguments{
\item{predictors}{Predictors are the known variables, based on which
we want to conduct our predicting.}

\item{outcome}{is the character string specifying the outcome variable.
This variable has to be a part of the database (db)}

\item{db}{The database on which we should perform the operations}
}
\value{
The table of non-missing observations after eliminating the variable
to the left.
}
\description{
This function helps you to decide which variable is restricting your analysis
by having the highest number of missing values in the selected set.
}
\examples{
which.restricts(c("d_age","b_sex"),"ANAscore",data)
}

